Added basic support for PxBatchQueryResult<HitType>.
I originally ignored this type because it was deprecated, but it turns out to still be necessary for using the vehicles module. In theory Biohazrd's template support should be able to instantiate the three variants of this type which are actually used, but right now template support is a little all or nothing so I manually translated the type for the sake of simplicity.