Added special case handling for PxFixedSizeLookupTable<8>.
This template is only ever used with 8 and only in a few places.
Ideally we'd like Biohazrd translate it automatically (and probably rename it to something more sensible), but Biohazrd currently chokes on some template stuff in PhysX and this was easier than either fixing it or implementing a filter.